+++ backend/board/src/main/java/com/github/jamie9504/board/role/RoleRequest.java
@@ -1,6 +1,4 @@
-package com.github.jamie9504.board.web.dto;
-
-import com.github.jamie9504.board.entity.Role;
+package com.github.jamie9504.board.role;
 
 public class RoleRequest {
