+++ backend/board/src/main/java/com/github/jamie9504/board/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.github.jamie9504.board.entity;
+package com.github.jamie9504.board.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
